feature empty crest dir #30
If for some reason the target dir exists and is empty, we could use
this. However this is not a trivial issue in bash and I just don't wanna
use the find to do it.

So we just delete it and recreate it. Dirty workaround ftw!

We'll still die with error when we can't delete the directory.