Release Version v0.5.0 (#35)
* Fix: #29 
Checks for the submission binaries creating proper error messages and fail.

* Feature: #30
If for some reason the target dir exists and is empty, we could use this. However this is not a trivial issue in bash without another program. Therefore, the script tries to delete and recreate the directory. It'll still die with error when it can't delete the directory.

* Fix: make debug more beautiful
It was not really necessary, but all lines of debug should be prepended, otherwise it's really hard to distinguish if it's just stdout when multilines are used.

* Feature (and fix): #31 
Openbabel is now optional.
It is no longer necessary to have obabel (this is the new default), it can be enabled via a commandline option though.
It is included in the config files and in the configure script.

* Fix: #32
Some guides could be adjusted, some need more work and only got the disclaimer.
All pdfs were removed. I'll need to rebuild those. Later.

* Minor improvements for clarity, change versions, change dates and cleanup or rogue spaces.
The script currently uses spaces for indentation.
I have converted.
This needs to be addressed in a later patch, because this is going to become messy. At least with all the changes here.